+++ src/errors.rs
@@ -9,14 +9,16 @@ pub enum BogrepError {
     CreateClient(reqwest::Error),
     #[error("Can't fetch website: {0}")]
     HttpResponse(reqwest::Error),
-    #[error("Invalid status code: {0}")]
-    HttpStatus(String),
+    #[error("Invalid status code ({url}): {status}")]
+    HttpStatus { status: String, url: String },
     #[error("Can't fetch website: {0}")]
     ParseHttpResponse(reqwest::Error),
-    #[error("Can't fetch binary bookmark")]
-    BinaryResponse,
-    #[error("Can't fetch empty bookmark")]
-    EmptyResponse,
+    #[error("Can't fetch binary bookmark ({0})")]
+    BinaryResponse(String),
+    #[error("Can't fetch empty bookmark ({0})")]
+    EmptyResponse(String),
+    #[error("Can't get host for url: {0}")]
+    ConvertHost(String),
     #[error("Can't serialize json: {0}")]
     SerializeJson(serde_json::Error),
     #[error("Can't deserialize json: {0}")]
@@ -25,8 +27,6 @@ pub enum BogrepError {
     ParseUrl(#[from] ParseError),
     #[error("Can't parse url: {0}")]
     ConvertHtml(readability::error::Error),
-    #[error("Can't get host for url: {0}")]
-    ConvertHost(String),
     #[error("Invalid utf8: {0}")]
     ConvertUtf8(#[from] FromUtf8Error),
     #[error("Can't read from HTML: {0}")]
@@ -34,27 +34,27 @@ pub enum BogrepError {
     #[error("Can't serialize HTML: {0}")]
     SerializeHtml(io::Error),
     #[error("Can't create file at {path}: {err}")]
-    CreateFile { path: String, err: String },
+    CreateFile { path: String, err: io::Error },
     #[error("Can't open file at {path}: {err}")]
-    OpenFile { path: String, err: String },
+    OpenFile { path: String, err: io::Error },
     #[error("Can't remove file at {path}: {err}")]
-    RemoveFile { path: String, err: String },
+    RemoveFile { path: String, err: io::Error },
     #[error("Can't read file: {0}")]
-    ReadFile(String),
+    ReadFile(io::Error),
     #[error("Can't write to file at {path}: {err}")]
-    WriteFile { path: String, err: String },
+    WriteFile { path: String, err: io::Error },
     #[error("Can't append file at {path}: {err}")]
-    AppendFile { path: String, err: String },
+    AppendFile { path: String, err: io::Error },
     #[error("Can't rename file from {from} to {to}: {err}")]
     RenameFile {
         from: String,
         to: String,
-        err: String,
+        err: io::Error,
     },
     #[error("Can't flush file: {0}")]
     FlushFile(io::Error),
     #[error("Can't rewind file: {0}")]
-    RewindFile(String),
+    RewindFile(io::Error),
     #[error("Can't convert header to string: {0}")]
     ConvertToStr(#[from] ToStrError),
     #[error("Can't remove website ({url}) from cache: {err}")]