+++ src/utils.rs
@@ -14,7 +14,7 @@ pub fn read_file(path: &Path) -> Result<Vec<u8>, BogrepError> {
     let mut buffer = Vec::new();
     let mut file = open_file(path)?;
     file.read_to_end(&mut buffer)
-        .map_err(|err| BogrepError::ReadFile(err.to_string()))?;
+        .map_err(BogrepError::ReadFile)?;
     Ok(buffer)
 }
 
@@ -25,7 +25,7 @@ pub fn read_file_to_string(path: &Path) -> Result<String, BogrepError> {
     let mut buffer = String::new();
     let mut file = open_file(path)?;
     file.read_to_string(&mut buffer)
-        .map_err(|err| BogrepError::ReadFile(err.to_string()))?;
+        .map_err(BogrepError::ReadFile)?;
     Ok(buffer)
 }
 
@@ -37,7 +37,7 @@ pub fn write_file(path: &Path, content: String) -> Result<(), BogrepError> {
     file.write_all(content.as_bytes())
         .map_err(|err| BogrepError::WriteFile {
             path: path.to_string_lossy().to_string(),
-            err: err.to_string(),
+            err,
         })?;
     file.flush().map_err(BogrepError::FlushFile)?;
     Ok(())
@@ -52,7 +52,7 @@ pub async fn write_file_async(path: &Path, content: &[u8]) -> Result<(), BogrepE
         .await
         .map_err(|err| BogrepError::WriteFile {
             path: path.to_string_lossy().to_string(),
-            err: err.to_string(),
+            err,
         })?;
     file.flush().await.map_err(BogrepError::FlushFile)?;
     Ok(())
@@ -64,7 +64,7 @@ pub fn open_file(path: &Path) -> Result<File, BogrepError> {
     debug!("Open file at {}", path.display());
     let file = File::open(path).map_err(|err| BogrepError::OpenFile {
         path: path.to_string_lossy().to_string(),
-        err: err.to_string(),
+        err,
     })?;
     Ok(file)
 }
@@ -77,7 +77,7 @@ pub fn open_file_in_read_mode(path: &Path) -> Result<File, BogrepError> {
         .open(path)
         .map_err(|err| BogrepError::OpenFile {
             path: path.to_string_lossy().to_string(),
-            err: err.to_string(),
+            err,
         })?;
     Ok(file)
 }
@@ -91,7 +91,7 @@ pub fn open_file_in_read_write_mode(path: &Path) -> Result<File, BogrepError> {
         .open(path)
         .map_err(|err| BogrepError::OpenFile {
             path: path.to_string_lossy().to_string(),
-            err: err.to_string(),
+            err,
         })?;
     Ok(file)
 }
@@ -106,7 +106,7 @@ pub fn open_and_truncate_file(path: &Path) -> Result<File, BogrepError> {
         .open(path)
         .map_err(|err| BogrepError::OpenFile {
             path: path.to_string_lossy().to_string(),
-            err: err.to_string(),
+            err,
         })?;
     Ok(file)
 }
@@ -116,7 +116,7 @@ pub fn create_file(path: &Path) -> Result<File, BogrepError> {
     debug!("Create file at {}", path.display());
     let file = File::create(path).map_err(|err| BogrepError::CreateFile {
         path: path.to_string_lossy().to_string(),
-        err: err.to_string(),
+        err,
     })?;
     Ok(file)
 }
@@ -128,7 +128,7 @@ pub async fn create_file_async(path: &Path) -> Result<tokio::fs::File, BogrepErr
         .await
         .map_err(|err| BogrepError::CreateFile {
             path: path.to_string_lossy().to_string(),
-            err: err.to_string(),
+            err,
         })?;
     Ok(file)
 }
@@ -142,7 +142,7 @@ pub fn append_file(path: &Path) -> Result<File, BogrepError> {
         .open(path)
         .map_err(|err| BogrepError::AppendFile {
             path: path.to_string_lossy().to_string(),
-            err: err.to_string(),
+            err,
         })?;
     Ok(file)
 }
@@ -152,7 +152,7 @@ pub fn remove_file(path: &Path) -> Result<(), BogrepError> {
     debug!("Remove file at {}", path.display());
     fs::remove_file(path).map_err(|err| BogrepError::RemoveFile {
         path: path.to_string_lossy().to_string(),
-        err: err.to_string(),
+        err,
     })
 }
 
@@ -163,7 +163,7 @@ pub async fn remove_file_async(path: &Path) -> Result<(), BogrepError> {
         .await
         .map_err(|err| BogrepError::RemoveFile {
             path: path.to_string_lossy().to_string(),
-            err: err.to_string(),
+            err,
         })
 }
 
@@ -183,7 +183,7 @@ pub fn close_and_rename(from: (File, &Path), to: (File, &Path)) -> Result<(), Bo
     fs::rename(from.1, to.1).map_err(|err| BogrepError::RenameFile {
         from: from.1.to_string_lossy().to_string(),
         to: to.1.to_string_lossy().to_string(),
-        err: err.to_string(),
+        err,
     })?;
 
     Ok(())