+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   - Implement `bogrep -w <pattern>` (match only whole words)
 - changed
   - Set default for `max_idle_connections_per_host` from 100 to 10
+  - Fetch: Degrade hard failure to warning message for `BogrepError::CreateFile`
+    and `BogrepError::ConvertHost`
 
 ### v0.5.0